% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{calc_paper_similarity}
\alias{calc_paper_similarity}
\title{Calculate paper similarity scores}
\usage{
calc_paper_similarity(res, .f = mean)
}
\arguments{
\item{res}{A data frame containing paper comparison results with columns `paper1`, `paper2`, and `dist`.}

\item{.f}{Optional. A function to aggregate distances, defaults to `mean`.}
}
\value{
A data frame containing pairwise paper similarity scores, with columns `paper1`,
`paper2`, and `dist`. Papers with no comparison data are assigned a distance of 1.
}
\description{
Calculate paper similarity scores
}
