% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{clean_md}
\alias{clean_md}
\title{Extract JSON decisions from the markdown file}
\usage{
clean_md(file)
}
\arguments{
\item{file}{A single string containing a path to a JSON file}
}
\value{
A tibble containing the "decisions" field from the JSON data.
}
\description{
Extract the json block from the markdown file and convert it to a tibble.
}
