% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{count_paper_decisions}
\alias{count_paper_decisions}
\title{Count paper decisions}
\usage{
count_paper_decisions(df, items, sort = TRUE)
}
\arguments{
\item{df}{A data frame.}

\item{items}{A vector of column names.}

\item{sort}{Optional. A logical indicating whether to sort by count in descending order.}
}
\value{
A data frame with columns 'paper' and 'count'.
}
\description{
Count paper decisions
}
