% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{summarize_pdf}
\alias{summarize_pdf}
\title{Deprecated functions}
\usage{
summarize_pdf(...)

summarize_pdf(...)
}
\description{
`r lifecycle::badge("deprecated")`
* [summarize_pdf()]/ [summarise_pdf()] are renamed to [extract_decisions()].
}
\keyword{internal}
