% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose.R
\name{diag_decision_ppp}
\alias{diag_decision_ppp}
\alias{view_pairs}
\title{Diagnostics and inspection functions for paper similarity calculation}
\usage{
diag_decision_ppp(decision_df, distance_df)

view_pairs(paper_df, decision_df, pp1, pp2)
}
\arguments{
\item{decision_df}{A data frame containing item-level distances}

\item{distance_df}{Output from `calc_paper_similarity`}

\item{paper_df}{A data frame containing paper codings}

\item{pp1, pp2}{A single string each to identify the two papers}
}
\value{
Diagnostic summary

A data frame containing decision-level comparisons between the two papers.
}
\description{
Diagnostics and inspection functions for paper similarity calculation

View the coding detail for a pair of papers
}
