% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dossier-package.R
\docType{package}
\name{dossier-package}
\alias{dossier}
\alias{dossier-package}
\title{dossier: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/dossier}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/dossier/issues}
}

}
\keyword{internal}
