% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{edit_decisions}
\alias{edit_decisions}
\title{Edit Decisions in Shiny App}
\usage{
edit_decisions()
}
\value{
No return value, called for side effects.
}
\description{
This function launches a Shiny app that allows users to edit decisions.
}
