% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{filter_papers}
\alias{filter_papers}
\alias{filter_var_type}
\title{Subset the decision table by the most common variable-type and the paper}
\usage{
filter_papers(df, n = NULL, n_value = NULL)

filter_var_type(df, n = NULL, n_value = NULL)
}
\arguments{
\item{df}{The decision table object}

\item{n}{Optional. A numeric value indicating the number of papers to keep.}

\item{n_value}{Optional. A numeric value indicating the minimum number of decisions a paper must have to be kept.}
}
\description{
To ensure a reasonable computation on paper similarity, [filter_var_type()]
allows you to subset the decision table by the most
common variable/ type combinations among all papers. [filter_papers()]
allows you to subset the paper with at least x decisions or the top x papers
with the most decisions.
}
\examples{
raw_df <- read.csv(system.file("papers.csv", package = "dossier")) |> tibble::as_tibble()
tbl_df <- as_decision_tbl(raw_df)
df <- tbl_df |> filter_var_type(n = 6)
df2 <- tbl_df |> filter_var_type(n_value = 3)
identical(df, df2)
}
