% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{summarize_num_decisions_pp}
\alias{summarize_num_decisions_pp}
\alias{filter_papers}
\alias{summarize_decisions_ppp}
\alias{filter_var_type}
\alias{count_variable_type}
\alias{create_good_variables}
\alias{summarize_variable_binary}
\title{Summarize number of decisions per paper}
\usage{
summarize_num_decisions_pp(df, sort = TRUE)

filter_papers(df, n_paper = NULL, n_count = NULL)

summarize_decisions_ppp(paper_df)

filter_var_type(df, n = NULL, n_value = NULL)

count_variable_type(df)

create_good_variables(df, n = NULL, n_value = NULL)

summarize_variable_binary(df_wide)
}
\arguments{
\item{df}{A data frame.}

\item{sort}{Optional. A logical indicating whether to sort results by count in descending order.}

\item{n_paper}{Optional. A numeric value.}

\item{n_count}{Optional. A numeric value.}

\item{paper_df}{A data frame containing paper codings}

\item{n}{Optional. A numeric value.}

\item{n_value}{Optional. A numeric value.}

\item{df_wide}{A data frame.}
}
\value{
A data frame containing papers and their decision counts.
}
\description{
Summarize number of decisions per paper
}
