% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{to_dist_mtx}
\alias{to_dist_mtx}
\title{Convert a dataframe to a distance matrix}
\usage{
to_dist_mtx(df, paper_cols = c("paper1", "paper2"))
}
\arguments{
\item{df}{A dataframe.}

\item{paper_cols}{Optional. A character vector of length 2 specifying the column names for the papers being compared. Default is `c("paper1", "paper2")`.}
}
\value{
A distance matrix object of class `"dist"`.
}
\description{
Convert a dataframe to a distance matrix
}
