% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{calc_item_similarity}
\alias{calc_item_similarity}
\alias{compute_text_embed}
\title{Calculate item similarity}
\usage{
calc_item_similarity(df, embed = NULL, text_model = "bert-base-uncased", ...)

compute_text_embed(df, text_model = "bert-base-uncased", ...)
}
\arguments{
\item{df}{A data frame.}

\item{embed}{Optional. A text embedding.}

\item{text_model}{A text model.}

\item{...}{Additional arguments passed to `text::textEmbed()`.}
}
\value{
A list of similarity scores between pairs of items.
}
\description{
Calculate item similarity
}
