% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{to_decision_tbl}
\alias{to_decision_tbl}
\title{Extract JSON decisions from the markdown file}
\usage{
to_decision_tbl(df)
}
\arguments{
\item{df}{A data frame or tibble to be casted into a decision table object}
}
\value{
A tibble of class decision_tbl
}
\description{
Extract the json block from the markdown file and convert it to a tibble.
}
\examples{
raw_df <- read.csv(system.file("papers.csv", package = "dossier")) |> tibble::as_tibble()
to_decision_tbl(raw_df)
}
