% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dossier-package.R
\docType{package}
\name{dossier-package}
\alias{dossier}
\alias{dossier-package}
\title{dossier: Reveal the Forking Paths of Decisions in Data Analysis}
\description{
Tools to extract and summarize decisions in the data analysis from PDF research papers. Methods to measure similarities between studies, cluster them, and visualize the results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/huizezhang-sherry/dossier}
  \item Report bugs at \url{https://github.com/huizezhang-sherry/dossier/issues}
}

}
\author{
\strong{Maintainer}: H. Sherry Zhang \email{huizezhangsh@gmail.com}

Authors:
\itemize{
  \item Roger Peng
}

}
\keyword{internal}
