% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{count_variable_type}
\alias{count_variable_type}
\alias{create_good_variables}
\alias{summarize_variable_binary}
\alias{summarize_num_decisions_pp}
\alias{slice_papers}
\alias{filter_variable_type}
\title{Summarize number of decisions per paper}
\usage{
count_variable_type(df)

create_good_variables(df, n = NULL, n_value = NULL)

summarize_variable_binary(df_wide)

summarize_num_decisions_pp(df_wide, sort = TRUE)

slice_papers(df, n_paper = NULL, n_count = NULL)

filter_variable_type(df, n = NULL, n_value = NULL)
}
\arguments{
\item{df}{A data frame.}

\item{n}{Optional. A numeric value.}

\item{n_value}{Optional. A numeric value.}

\item{df_wide}{A data frame.}

\item{sort}{Optional. A logical indicating whether to sort results by count in descending order.}

\item{n_paper}{Optional. A numeric value.}

\item{n_count}{Optional. A numeric value.}
}
\value{
A data frame containing papers and their decision counts.
}
\description{
Summarize number of decisions per paper
}
