% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{count_paper_decisions}
\alias{count_paper_decisions}
\alias{count_paper_pair_decisions}
\alias{count_variable_type}
\alias{skim_variable_binary}
\title{Small functions to summarize the decision table for pre-processing}
\usage{
count_paper_decisions(df, sort = TRUE)

count_paper_pair_decisions(df)

count_variable_type(df, sort = TRUE)

skim_variable_binary(df)
}
\arguments{
\item{df}{The decision table object.}

\item{sort}{A logical indicating whether to sort results by count in
descending order.Default to `TRUE`}
}
\value{
A data frame containing papers and their decision counts.
}
\description{
Small functions to summarize the decision table for pre-processing
}
\examples{
raw_df <- read.csv(system.file("papers.csv", package = "dossier")) |> tibble::as_tibble()
df <- as_decision_tbl(raw_df)
count_paper_decisions(df)
count_variable_type(df)
df |> filter_var_type(n = 6) |> count_paper_pair_decisions()
}
