% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-similarity.R
\name{to_dist_mtx}
\alias{to_dist_mtx}
\title{Convert distance vector to distance matrix}
\usage{
to_dist_mtx(distance_df, paper_df)
}
\arguments{
\item{distance_df}{A data frame containing a `dist` column.}

\item{paper_df}{A data frame containing a `paper` column.}
}
\value{
A distance matrix of class `"dist"` with labels from `paper_df$paper`.
}
\description{
Convert distance vector to distance matrix
}
